Added graph components and logic

+++ src/components/ExpenseItem/ExpenseItem.component.jsx
@@ -3,7 +3,7 @@ import { Card } from '../Card/Card.component'
 
 const ExpenseItem = (props) =>{
     
-    const month = props.initData.date.toLocaleString('en-US', {month: 'long'});
+    const month = props.initData.date.toLocaleString('en-US', {month: 'short'});
     const day = props.initData.date.toLocaleString('en-US', {day: '2-digit'});
     const year = props.initData.date.getFullYear();
